refactor: separate DigitalClockwork into its own header and source files
