+++ Makefile
@@ -1,15 +1,15 @@
 CXX = g++
-CXXFLAGS = -std=c++17 -Wall -Wextra -Iinclude
+CXXFLAGS = -std=c++17 -Wall -Wextra -Iinclude $(shell pkg-config --cflags libevdev)
 
+LIBS = $(shell pkg-config --libs libevdev)
 SRC = src/chips.cpp src/freqGenerator.cpp src/feedback.cpp
+
 BUILD = builds
+TEST_NAMES = 4511 4029 4040 4017 4013 4063 frequency keyboard
 
-TEST_NAMES = 4511 4029 4040 4017 4063 4013 frequency
 TEST_BINS = $(addprefix $(BUILD)/, $(addsuffix test, $(TEST_NAMES)))
-
-CLOCK_SRC = src/digitalClockwork.cpp
+CLOCK_SRC = src/main.cpp src/digitalClockwork.cpp
 CLOCK_BIN = $(BUILD)/digitalClock
-
 
 .PHONY: all tests clean runClock
 
@@ -18,15 +18,16 @@ all: tests
 $(BUILD):
 	mkdir -p $(BUILD)
 
-
 # TESTS
 tests: $(BUILD) $(TEST_BINS)
 
 
-# Regra genérica para todos os testes
 $(BUILD)/%test: tests/%test.cpp $(SRC)
-	$(CXX) $(CXXFLAGS) $< $(SRC) -o $@
+	$(CXX) $(CXXFLAGS) $< $(SRC) -o $@ $(LIBS)
 
+
+$(BUILD)/keyboardtest: tests/keyboardtest.cpp
+	$(CXX) $(CXXFLAGS) $< -o $@ $(LIBS)
 
 
 # CLOCKWORK
@@ -34,9 +35,8 @@ runClock: $(BUILD) $(CLOCK_BIN)
 	./$(CLOCK_BIN)
 
 
-
 $(CLOCK_BIN): $(CLOCK_SRC) $(SRC)
-	$(CXX) $(CXXFLAGS) $^ -o $@
+	$(CXX) $(CXXFLAGS) $^ -o $@ $(LIBS)
 
 
 # CLEAR
