+++ src/main.cpp
@@ -1,0 +1,181 @@
+/*
+[PT-BR]
+    O Digital Clockwork simula a lógica do circuito de um relógio digital inspirado no projeto de Wagner Rambo (WR Kits). 
+    A classe DigitalClockwork orquestra a interação entre contadores BCD (CD4029), decodificadores de sete segmentos (CD4511), 
+    portas AND (CD4081), um contador Johnson (CD4017), um contador de ondulação (CD4040) e um gerador de frequência.
+
+    Cada chip foi modelado individualmente em uma classe separada, declarada em chips.hpp e implementada em chips.cpp. O gerador 
+    de frequência foi modelado com o uso de threads para maior precisão temporal e, assim como os chips, foi encapsulado em sua 
+    própria classe, declarada em freqGenerator.hpp e implementada em freqGenerator.cpp. Os arquivos feedback.hpp e feedback.cpp
+    foram usados, respectivamente, para a declaração e implementação dos artifícios responsáveis pelo feedback visual do usuário 
+    no terminal.
+
+[EN-US]
+    A Digital Clockwork simulates the logic of a digital clock circuit inspired by the project of Wagner Rambo (WR Kits).
+    The class "DigitalClockwork" orchestrates the interaction between BCD counters (CD4029), seven-segment decoders (CD4511), 
+    AND gates (CD4081), a Johnson counter (CD4017), a ripple counter (CD4040), and a frequency generator.
+
+    Each chip was individually modeled in a separate class, declared in *chips.hpp* and implemented in *chips.cpp*. The frequency 
+    generator was modeled using threads to achieve greater timing precision and, like the chips, was encapsulated in its own class, 
+    declared in *freqGenerator.hpp* and implemented in *freqGenerator.cpp*. The files *feedback.hpp* and *feedback.cpp* were used, 
+    respectively, for the declaration and implementation of the mechanisms responsible for providing visual feedback to the user in 
+    the terminal.
+*/
+
+//libraries
+#include "digitalClockwork.hpp"                         // The Wagner Rambo's original project simulation
+#include "digitalAlarm.hpp"                             // the extension: The alarm system for The Digital Clockwork
+
+#include "chips.hpp"                                    // CD4029, CD4511, CD4081, CD4017, CD4040...
+#include "feedback.hpp"                                 // Display, Led...
+#include "freqGenerator.hpp"                            // Frequency generator
+
+#include <iostream>                                     // std::cout, std::cerr
+#include <cstdlib>                                      // system()
+#include <string>                                       // std::string
+#include <array>                                        // std::array
+
+#include <libevdev/libevdev.h>                          // keyboard input events
+#include <fcntl.h>                                      // open(), O_RDONLY, O_NONBLOCK
+#include <unistd.h>                                     // close()
+
+
+
+int main(){
+    // open keyboard device
+    int fd = open("/dev/input/event4", O_RDONLY | O_NONBLOCK);
+
+    if(fd < 0){
+        std::cerr << "The file couldn't open\n";
+        return EXIT_FAILURE;
+    }
+
+    // initialize libevdev device from file descriptor
+    struct libevdev* dev = nullptr;
+    libevdev_new_from_fd(fd, &dev);
+
+
+
+    // initial preset value for all BCD counters
+    const std::array<bool, 4> presetZero = {0,0,0,0};
+
+    
+    // instantiate BCD counters, decoders and AND gate chips
+    std::array<Chip4029, 4> cd4029 = {
+        Chip4029(presetZero),
+        Chip4029(presetZero),
+        Chip4029({1,0,0,0}),
+        Chip4029(presetZero)
+    };
+
+    std::array<Chip4511, 4> cd4511{};
+    std::array<Chip4081, 2> cd4081{};  
+                                                        
+    FreqGenerator clk;                                  // frequency: 60 Hz       
+    Chip4017 cd4017(2, true);                           // AM/PM indicator
+    Chip4040 cd4040;                                    // 12-stage ripple counter — frequency divider
+
+
+    // feedback components
+    Display display;
+    Led AM;
+    Led PM;
+
+    
+    // create pointer arrays to pass chips into DigitalClockwork
+    std::array<Chip4029*, 4> ptr4029;
+    std::array<Chip4511*, 4> ptr4511;
+    std::array<Chip4081*, 2> ptr4081;
+
+    for(int i = 0; i < 4; i++){
+        ptr4029[i] = &cd4029[i];
+        ptr4511[i] = &cd4511[i];
+    }
+
+    for(int i = 0; i < 2; i++){
+        ptr4081[i] = &cd4081[i];
+    }
+
+
+    // instantiate the main clockwork controller
+    DigitalClockwork functions(
+        ptr4029,
+        ptr4511,
+        ptr4081,
+        &cd4017,
+        &cd4040  
+    );
+
+
+    // configure all counters and decoders             
+    for(int i = 0; i < 4; i++){
+        cd4511[i].setLampTest(true);                    // enable normal display output
+        cd4511[i].setBlanking(true);                    // enable segment output
+        cd4029[i].setBinaryDecade(false);               // set to decimal (BCD) mode
+        cd4029[i].setUpDown(true);                      // set to count up
+    }
+
+
+    // start the 60Hz frequency generator
+    clk.start();
+    bool lastState = clk.getState();
+
+    DigitalClockwork::ADJUSTMENT mode = DigitalClockwork::ADJUSTMENT::DEFAULT;
+
+    // updates the clock on each rising edge of the frequency generator
+    while(true) {
+        // read keyboard input
+        struct input_event ev;
+
+        if(libevdev_next_event(dev, LIBEVDEV_READ_FLAG_NORMAL, &ev) == 0){
+            if(ev.type == EV_KEY){
+                // key pressed
+                if(ev.value == 1){
+                    if(ev.code == KEY_F){
+                        mode = DigitalClockwork::ADJUSTMENT::FAST;
+                    }
+
+                    if(ev.code == KEY_S){
+                        mode = DigitalClockwork::ADJUSTMENT::SLOW;
+                    }
+                }
+
+                // key released
+                if(ev.value == 0){
+                    if(ev.code == KEY_F || ev.code == KEY_S){
+                        mode = DigitalClockwork::ADJUSTMENT::DEFAULT;
+                    }
+                }
+            }
+        }
+
+        
+        // wait for rising edge and update the circuit
+        clk.waitEdge(lastState);
+        bool curState = clk.getState();
+
+        if(!lastState && curState){
+            system("clear");
+            functions.updateSystem(mode);
+
+            std::array<std::array<bool, 7>, 4> segments = functions.getSegmentsOutput();
+
+            AM.setState(functions.getMeridien(0));
+            PM.setState(functions.getMeridien(1));
+
+            std::cout << display.render(segments) << " | "; 
+            std::cout << "AM:" << AM.getState() << "  PM:" << PM.getState() << " |\n";        
+        }
+        
+        lastState = curState;
+    }
+
+
+    // stop the frequency generator, the keyboard reader and finish this program
+    clk.stop();
+
+    libevdev_free(dev);
+    close(fd);
+
+    return EXIT_SUCCESS;
+}
